Moves everything from container CheckoutDisplay into component SummaryBox for consistency
